Merge pull request #9 from hogenshpogen/TAMU_gtxrx_bb_hotch
Fiber reset, Bad bits, Hot channels

Some features are not quite working because functions that are supposed to be doing it depend on the hardware version value,
which is not set at the moment of the calls. This will be fixed separately (planned already from earlier discussions with Evaldas and Jinghua).

Among those affected are:
- read_hot_channel_mask_  values are not showing up correctly in PrintConfiguration,
which is likely related to  hcm* addresses not making it to the list of the config registers

-------------------------------------------
TMB Hot Channel Mask (from right to left)
-------------------------------------------
Layer 5 -> 00 00 ff ff ff ff ff
Layer 4 -> 00 00 ff ff ff ff ff
Layer 3 -> 00 00 ff ff ff ff ff
Layer 2 -> 00 00 ff ff ff ff ff
Layer 1 -> 00 00 ff ff ff ff ff
Layer 0 -> 00 00 ff ff ff ff ff

- dcfeb_badbits_ctrl_adr is not on the list of the configuration registers
(as is, it's an omission, but even if added with hw version check, it will not show up because the version isn't there yet).